+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
 
 ## 2025-03-08
+- UIKit から SwiftUI ベースに変更
+- 「ハロー」「コンニチワ」メッセージを表示する機能を実装
 - docs/tasks.md ファイルを作成（GitHub の TODO 形式で作業リストを記載）